:sparkles: [feat] 포포 스테이지에서 end 버튼 누르면 스켈레톤 배열 서버로 보내고 값 얻기 기능 추가

+++ lib/domain/provider/popo_skeleton_provider.dart
@@ -1,3 +1,3 @@
 abstract class PoPoSkeletonProvider {
-  Future postSkeletonList(List<List<int>> seq);
+  Future postSkeletonList(List<List<double>> seq);
 }
